+++ crates/aioduct/src/upgrade/mod.rs
@@ -5,5 +5,5 @@ mod upgrade_send;
 
 pub use upgrade_local::UpgradedLocal;
 pub(crate) use upgrade_local::{UpgradeHandleLocal, on_upgrade_local_manual};
-pub use upgrade_send::Upgraded;
+pub use upgrade_send::UpgradedSend;
 pub(crate) use upgrade_send::on_upgrade;